# PERF: Series.duplicated for pyarrow timestamp and duration types