--- a/asv_bench/benchmarks/algorithms.py
+++ b/asv_bench/benchmarks/algorithms.py
@@ -96,8 +96,12 @@ class Duplicated:
             "datetime64[ns, tz]": pd.date_range(
                 "2011-01-01", freq="H", periods=N, tz="Asia/Tokyo"
             ),
-            "timestamp[ms][pyarrow]": pd.Index(np.arange(N), dtype=dtype),
-            "duration[s][pyarrow]": pd.Index(np.arange(N), dtype=dtype),
+            "timestamp[ms][pyarrow]": pd.Index(
+                np.arange(N), dtype="timestamp[ms][pyarrow]"
+            ),
+            "duration[s][pyarrow]": pd.Index(
+                np.arange(N), dtype="duration[s][pyarrow]"
+            ),
         }[dtype]
         if not unique:
             data = data.repeat(5)