--- a/pandas/core/arrays/base.py
+++ b/pandas/core/arrays/base.py
@@ -61,7 +61,7 @@ from pandas.core import (
     roperator,
 )
 from pandas.core.algorithms import (
-    duplicated_array,
+    duplicated,
     factorize_array,
     isin,
     map_array,
@@ -1140,7 +1140,7 @@ class ExtensionArray:
         >>> pd.array([1, 1, 2, 3, 3], dtype="Int64").duplicated()
         array([False,  True, False, False,  True])
         """
-        return duplicated_array(values=self, keep=keep)
+        return duplicated(values=self, keep=keep)
 
     def shift(self, periods: int = 1, fill_value: object = None) -> ExtensionArray:
         """