- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.

Alternatively, we could define `ExtensionArray._duplicated` to avoid accessing private methods from outside the class. 

`asv continuous -f 1.1 upstream/main temp-dup -b algorithms.Duplicated`

```
       before           after         ratio
-      1.97±0.06s       22.9±0.4ms     0.01  algorithms.Duplicated.time_duplicated(False, False, 'timestamp[ms][pyarrow]')
-      1.93±0.01s       17.7±0.6ms     0.01  algorithms.Duplicated.time_duplicated(False, 'last', 'timestamp[ms][pyarrow]')
-      1.98±0.07s         17.7±2ms     0.01  algorithms.Duplicated.time_duplicated(False, 'first', 'timestamp[ms][pyarrow]')
-       4.43±0.1s       22.7±0.2ms     0.01  algorithms.Duplicated.time_duplicated(False, False, 'duration[s][pyarrow]')
-       4.01±0.1s       17.5±0.6ms     0.00  algorithms.Duplicated.time_duplicated(False, 'first', 'duration[s][pyarrow]')
-      4.05±0.06s       17.1±0.5ms     0.00  algorithms.Duplicated.time_duplicated(False, 'last', 'duration[s][pyarrow]')
```