@rhshadrach 2023-10-20
This patch may have induced a performance regression. If it was a necessary behavior change, this may have been expected and everything is okay.

Please check the links below. If any ASVs are parameterized, the combinations of parameters that a regression has been detected for appear as subbullets.

 - [period.Algorithms.time_drop_duplicates](https://asv-runner.github.io/asv-collection/pandas/#period.Algorithms.time_drop_duplicates)
   - [typ='index'](https://asv-runner.github.io/asv-collection/pandas/#period.Algorithms.time_drop_duplicates?p-typ=%27index%27)

Subsequent benchmarks may have skipped some commits. The link below lists the commits that are between the two benchmark runs where the regression was identified.

https://github.com/pandas-dev/pandas/compare/6e6a6831aaa2d1606aa897b34f106e655180034b...3bf0f647ca42bf21de76f9e6027de38cb5589147